+++ app/controllers/selection_statuses_controller.rb
@@ -38,12 +38,17 @@ class SelectionStatusesController < ApplicationController
     @selection_status_destroy = SelectionStatus.find(params[:id])
     @selection_statuses = SelectionStatus.where(user_id:current_user.id).order(id: "DESC")
     @selection_status = SelectionStatus.new
-    if @selection_status_destroy.destroy
-      flash[:notice] = "削除に成功しました"
-      redirect_to selection_statuses_path
-    else
-      flash[:alert] = "削除に失敗しました"
+    if @selection_status_destroy.company.count >= 1
+      flash[:alert] = "この選考状況カテゴリは使用されているので削除できません"
       render 'index'
+    else
+      if @selection_status_destroy.destroy
+        flash[:notice] = "削除に成功しました"
+        redirect_to selection_statuses_path
+      else
+        flash[:alert] = "削除に失敗しました"
+        render 'index'
+      end
     end
   end
